Fix: fixed issue that led to negative kinetic parameters
This was done by taking into account the sign of
the reaction flux when calculating the reversibilities,
as in the GRASP paper.
However, because inhibition steps have no reversibilities
(they are at equilibrium), modifierElemflux in
caculateKineticParams was also multiplied by the
reaction flux sign so that all rate constants are
positive.