Fix: added assertion to check for incompatible fluxes and Gibbs energies
It is possible that the mean reference flux for a given reaction is
incompatible with the respective Gibbs energies if the flux standard
deviation is high enough to allow the reaction to go in the opposite
direction.
In this case TMFA is able to find feasible flux values, but because
the mean reference flux is used to calculate the reversibilities
and that is incompatible with the Gibbs energy, it will lead to
negative kinetic parameters.