Feat: optimized memory usage when importing simulations
Previously importing and processing simulation
results took a huge ammount of memory, making
it hard to process simulations for 10.000 models
with 50+ reactions on a laptop with 16GB of RAM.
Now we use pre-allocated numpy arrays to import
the data instead of continuously extending lists,
and have an option to import only medians/IQRs
for concentrations and fluxes independently, so
that the notebook stays small as well.
Now there are 2 notebooks: one for small
simulations and another for large ones.