Feat: removed dependency on mets active and fixed cols
metsActive are now metsBalanced if only reference
state or one condition; they are mets that change
across conditions if more than one condition.
metsFixed are the complement of metsBalanced
if only reference state or one condition; they
are the mets that do not change over conditions
if more than one condition.