tests: Update test comparison for numerical values
For tests with numerical results, test that the
true and actual results are the same within
a defined relative tolerance OR absolute tolerance.
The idea is that for large values the relative tolerance
will be the one that matters, and for smaller values it
will be the absolute tolerance that matters.
If both tolerances are very low it should work.