Update buildEnsemble.m
This is a somewhat major modification for efficient parallelization of the code. Briefly, the Parallel Toolbox of MATLAB contains functions that allow the execution of parallel tasks with communication between the workers, as opposed to parfor(). One of those functions is spmd() (Single Program Multiple Data). Handling the data is a bit tricky as it is now contained in 'Composites', however, I managed to make it work. By using this script, now we can fully exploit parallel computation, and, at the same time, we can run more complicated loops where we would like to set several conditions for termination.
Finally, I had to make some changes to the builEnsemble function for compatibility issues.